requirements.txt
streamlit
spacy
pdfminer.six
python-docx
sentence-transformers
pytesseract
Pillow
plotlypip